+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to Hourly Sensor will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v0.3.0] - 2026-08-18
+
+### Added
+
+- Add a button for every configured hourly sensor that forces its active window
+  to be recalculated from the source entity's Recorder history.
+
 ## [v0.2.2] - 2026-08-02
 
 ### Fixed